+++ webapp/src/router/index.ts
@@ -144,8 +144,8 @@ const routes: Array<RouteRecordRaw> = [
     },
   },
   {
-    path: '/greentreasury',
-    name: 'greentreasury',
+    path: '/energytreasury',
+    name: 'energytreasury',
     component: () => import('@/views/GreenTreasuryView.vue'),
       meta: {
       requiresNotMainNetwork: true